requirements.txt
Biopython==1.68
pysam==0.9.1.4
coverage==4.2
jenkspy==0.1.3
pyutilsnrw>=0.0.768
numpy>=1.11.2
matplotlib==1.5.3
pandas==0.18.1
git+https://github.com/WarrenWeckesser/heatmapcluster.git#egg=heatmapcluster
